Wrote concepts; built screens structure
The FeedScreen is in the making, using generated data from the new
DataGenerator. A new yarn package 'moment' was added to handle
displaying of dates in a simple manner. Social App concepts document was
used to structure the data itself.

Along with that, there were some hidden files forgotten to be included
in the repository.