Persistent local storage now being used in FeedScreen.
Added data initializer to first screen in tab nav; adjusted
DataGenerator for use with this.
PostingView can now add posts to top of feed in FeedScreen.
Generating a post now allows a given percent-chance to reserve a user id
for it. A similar chance is given to assigning a parent post id.
Generating conversations may now accept any given array of users to talk
to. A target user is selected by chance, and may be selected more than
once.
Exported ID_LIMITs in DataGenerator in case of needed dummy data
generation.
Renamed WriteView to WritingView.