User can now post replies.
Replies and new posts are now visible on navigation between home tabs,
and in post view.
There is a bug for trying to send a reply; it requires a second click to
activate the reply action.
Adjusted the data initializer to avoid duplicate userIds.
Messages can now access conversation threads; however, it cannot
currently reply to them.