Messaging, following and favoriting/liking are now implemented.
Refactored data structure of messages for simplified adding, at the cost
of slower fetching of them.
The conversation view now uses GiftedChat for convenience. An unused
file containing the older incomplete implementation is kept.
The writing message view uses a third-party component for searching
possible people to message. That search includes only people the user is
following.
Fixed Feed bug that cleared out all other posts when posting or
replying.
Objects and arrays directly passed to some screens are now converted to
JSON strings and back.
Removed extra padding in collapsible reply view.
Moved the 'clear data' functionality to its own dedicated button in the
profile page. That function was originally put in the search button in
the feed page.
Added a copy of the app.