Started building functionality for the three stats in ProfileScreen.
Added the display of stats in the StatsTabs navigator.
Moved the ProfileScreen's promises outside their useEffect functions, to
shorten code. Data fetching is also triggered via navigation event
instead.
Added interactivity for FeedScreen's floating buttons. The blank space
in the list is a header to be filled with a new post textbox.